backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
pydantic[email]>=2.6.3
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
google-generativeai>=0.3.2
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
sqlalchemy>=2.0.28
alembic>=1.13.1
aiosqlite>=0.20.0
email-validator
